Extract shared domain models for the allocation management
