+++ allocations/access-amie/go.mod
@@ -3,11 +3,13 @@ module github.com/apache/airavata-custos/allocations/access-amie
 go 1.22.0
 
 require (
+	github.com/apache/airavata-custos/allocations/domain v0.0.0
 	github.com/go-sql-driver/mysql v1.8.1
 	github.com/golang-migrate/migrate/v4 v4.18.1
 	github.com/google/uuid v1.6.0
 	github.com/jmoiron/sqlx v1.4.0
 	github.com/prometheus/client_golang v1.20.5
+	github.com/prometheus/client_model v0.6.1
 	github.com/stretchr/testify v1.10.0
 	google.golang.org/protobuf v1.36.4
 	gopkg.in/yaml.v3 v3.0.1
@@ -23,7 +25,6 @@ require (
 	github.com/klauspost/compress v1.17.9 // indirect
 	github.com/munnerz/goautoneg v0.0.0-20191010083416-a7dc8b61c822 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/prometheus/client_model v0.6.1 // indirect
 	github.com/prometheus/common v0.55.0 // indirect
 	github.com/prometheus/procfs v0.15.1 // indirect
 	github.com/stretchr/objx v0.5.2 // indirect
@@ -31,4 +32,7 @@ require (
 	golang.org/x/sys v0.25.0 // indirect
 )
 
-replace github.com/apache/airavata-custos/allocations/provisioner => ../provisioner
+replace (
+	github.com/apache/airavata-custos/allocations/domain => ../domain
+	github.com/apache/airavata-custos/allocations/provisioner => ../provisioner
+)
