+++ allocations/access-amie/handler/request_account_create_test.go
@@ -23,6 +23,7 @@ import (
 	"testing"
 
 	"github.com/apache/airavata-custos/allocations/access-amie/model"
+	dmodel "github.com/apache/airavata-custos/allocations/domain/model"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -34,42 +35,42 @@ import (
 
 type mockRACPersonService struct{ mock.Mock }
 
-func (m *mockRACPersonService) FindOrCreateFromPacket(ctx context.Context, tx *sql.Tx, body map[string]any) (*model.Person, error) {
+func (m *mockRACPersonService) FindOrCreateFromPacket(ctx context.Context, tx *sql.Tx, body map[string]any) (*dmodel.Person, error) {
 	args := m.Called(ctx, tx, body)
 	if args.Get(0) == nil {
 		return nil, args.Error(1)
 	}
-	return args.Get(0).(*model.Person), args.Error(1)
+	return args.Get(0).(*dmodel.Person), args.Error(1)
 }
 
 type mockRACAccountService struct{ mock.Mock }
 
-func (m *mockRACAccountService) ProvisionClusterAccount(ctx context.Context, tx *sql.Tx, person *model.Person) (*model.ClusterAccount, error) {
+func (m *mockRACAccountService) ProvisionClusterAccount(ctx context.Context, tx *sql.Tx, person *dmodel.Person) (*dmodel.ClusterAccount, error) {
 	args := m.Called(ctx, tx, person)
 	if args.Get(0) == nil {
 		return nil, args.Error(1)
 	}
-	return args.Get(0).(*model.ClusterAccount), args.Error(1)
+	return args.Get(0).(*dmodel.ClusterAccount), args.Error(1)
 }
 
 type mockRACProjectService struct{ mock.Mock }
 
-func (m *mockRACProjectService) CreateOrFindProject(ctx context.Context, tx *sql.Tx, projectID, grantNumber string) (*model.Project, error) {
+func (m *mockRACProjectService) CreateOrFindProject(ctx context.Context, tx *sql.Tx, projectID, grantNumber string) (*dmodel.Project, error) {
 	args := m.Called(ctx, tx, projectID, grantNumber)
 	if args.Get(0) == nil {
 		return nil, args.Error(1)
 	}
-	return args.Get(0).(*model.Project), args.Error(1)
+	return args.Get(0).(*dmodel.Project), args.Error(1)
 }
 
 type mockRACMembershipService struct{ mock.Mock }
 
-func (m *mockRACMembershipService) CreateMembership(ctx context.Context, tx *sql.Tx, projectID, clusterAccountID, role string) (*model.ProjectMembership, error) {
+func (m *mockRACMembershipService) CreateMembership(ctx context.Context, tx *sql.Tx, projectID, clusterAccountID, role string) (*dmodel.ProjectMembership, error) {
 	args := m.Called(ctx, tx, projectID, clusterAccountID, role)
 	if args.Get(0) == nil {
 		return nil, args.Error(1)
 	}
-	return args.Get(0).(*model.ProjectMembership), args.Error(1)
+	return args.Get(0).(*dmodel.ProjectMembership), args.Error(1)
 }
 
 type mockRACAmieClient struct{ mock.Mock }
@@ -110,9 +111,9 @@ func TestRequestAccountCreateHandler(t *testing.T) {
 			name:  "valid packet processes successfully",
 			input: validFixture,
 			setupMocks: func(ps *mockRACPersonService, as *mockRACAccountService, prj *mockRACProjectService, ms *mockRACMembershipService, ac *mockRACAmieClient, aud *mockRACAuditService) {
-				person := &model.Person{ID: "person-123"}
-				account := &model.ClusterAccount{ID: "account-123", Username: "testuser"}
-				project := &model.Project{ID: "test-project-456", GrantNumber: "TEST123"}
+				person := &dmodel.Person{ID: "person-123"}
+				account := &dmodel.ClusterAccount{ID: "account-123", Username: "testuser"}
+				project := &dmodel.Project{ID: "test-project-456", GrantNumber: "TEST123"}
 
 				ps.On("FindOrCreateFromPacket", mock.Anything, mock.Anything, mock.Anything).Return(person, nil)
 				aud.On("Log", mock.Anything, mock.Anything, mock.Anything, mock.Anything, model.AuditCreatePerson, mock.Anything, mock.Anything, mock.Anything).Return(nil)
@@ -122,7 +123,7 @@ func TestRequestAccountCreateHandler(t *testing.T) {
 
 				prj.On("CreateOrFindProject", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(project, nil)
 
-				ms.On("CreateMembership", mock.Anything, mock.Anything, "test-project-456", account.ID, "USER").Return(&model.ProjectMembership{}, nil)
+				ms.On("CreateMembership", mock.Anything, mock.Anything, "test-project-456", account.ID, "USER").Return(&dmodel.ProjectMembership{}, nil)
 				aud.On("Log", mock.Anything, mock.Anything, mock.Anything, mock.Anything, model.AuditCreateMembership, mock.Anything, mock.Anything, mock.Anything).Return(nil)
 
 				ac.On("ReplyToPacket", mock.Anything, int64(233497917), mock.Anything).Return(nil)
