+++ allocations/access-amie/service/person_service_test.go
@@ -23,7 +23,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/apache/airavata-custos/allocations/access-amie/model"
+	"github.com/apache/airavata-custos/allocations/domain/model"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
