+++ allocations/domain/go.mod
@@ -1,0 +1,5 @@
+module github.com/apache/airavata-custos/allocations/domain
+
+go 1.22.0
+
+require github.com/jmoiron/sqlx v1.4.0
