+++ allocations/domain/model/membership.go
@@ -1,0 +1,30 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package model
+
+import "time"
+
+type ProjectMembership struct {
+	ID               string    `db:"id" json:"id"`
+	ProjectID        string    `db:"project_id" json:"project_id"`
+	ClusterAccountID string    `db:"cluster_account_id" json:"cluster_account_id"`
+	Role             *string   `db:"role" json:"role,omitempty"`
+	IsActive         bool      `db:"is_active" json:"is_active"`
+	CreatedAt        time.Time `db:"created_at" json:"created_at"`
+	UpdatedAt        time.Time `db:"updated_at" json:"updated_at"`
+}
