+++ allocations/go.work
@@ -1,6 +1,7 @@
 go 1.22.0
 
 use (
+	./domain
 	./provisioner
 	./access-amie
 )
